End of Ep 11
Added layer and timer,  Optimized.

+++ Leopard/src/graphic/Buffer/VertexArray.cpp
@@ -11,7 +11,7 @@ namespace Leopard
 
 		VertexArray::~VertexArray()
 		{
-			for (int i = 0; i < m_Buffers.size(); i++)
+			for (size_t i = 0; i < m_Buffers.size(); i++)
 			{
 				delete m_Buffers[i];
 			}
